requirement_no_torch.txt
opencv-python
pandas
Pillow
protobuf
pycm
pydicom
scikit-image
scikit-learn
scipy
tensorboard
tensorboard-data-server
tensorboard-plugin-wit
timm
tqdm
simplejson
iopath
monai
regex
ftfy
braceexpand
huggingface_hub
sentencepiece
medmnist
vit_pytorch
torch-summary
fairscale
einops
h5py
transformers
